move Range code into range.go
